latest update

+++ AdminPortal/Controllers/TenantController.cs
@@ -1,4 +1,5 @@
 ﻿using Admin.Portal.API.Core.Const;
+using Admin.Portal.API.Core.Enum;
 using Admin.Portal.API.Core.Models;
 using Admin.Portal.API.Core.Models.Base;
 using Admin.Portal.API.Core.Request;
@@ -29,6 +30,9 @@ namespace Admin.Portal.API.Controllers
         [HttpGet, Route("Tenants")]
         public async Task<IActionResult> Tenants(int? userID)
         {
+            if(!new AuthenticateFilter(config, dbContext).Authenticate(AccessLevel.Admins, Claims.CLAIM_READ_TENANT, Request.Headers[Config.HEADER_LOGIN_USER].ToString()))
+                return new Context(Messages.ACCOUNT_INSUFFICIENT_PRIVILEGES).ToContextResult((int)HttpStatusCode.Forbidden);
+
             IDataAccess db = ServiceInit.GetDataInstance(config, dbContext);
             (bool, List<TenantModel>) result = await db.GetTenantDeatils(userID);
             if (result.Item1)
@@ -40,6 +44,9 @@ namespace Admin.Portal.API.Controllers
         [HttpPost, Route("Create")]
         public async Task<IActionResult> Create([FromBody] TenantRequest context)
         {
+            if (!new AuthenticateFilter(config, dbContext).Authenticate(AccessLevel.Admins, Claims.CLAIM_CREATE_TENANT, Request.Headers[Config.HEADER_LOGIN_USER].ToString()))
+                return new Context(Messages.ACCOUNT_INSUFFICIENT_PRIVILEGES).ToContextResult((int)HttpStatusCode.Forbidden);
+
             IDataAccess db = ServiceInit.GetDataInstance(config, dbContext);
             (bool, TenantModel) result = await db.CreateTenant(new TenantModel() { Name = context.Name});
             if (result.Item1)
@@ -49,9 +56,11 @@ namespace Admin.Portal.API.Controllers
         }
 
         [HttpPut, Route("Update")]
-
         public async Task<IActionResult> Update([FromBody] TenantModel context)
         {
+            if (!new AuthenticateFilter(config, dbContext).Authenticate(AccessLevel.Admins, Claims.CLAIM_UPDATE_TENANT, Request.Headers[Config.HEADER_LOGIN_USER].ToString()))
+                return new Context(Messages.ACCOUNT_INSUFFICIENT_PRIVILEGES).ToContextResult((int)HttpStatusCode.Forbidden);
+
             IDataAccess db = ServiceInit.GetDataInstance(config, dbContext);
             (bool, TenantModel) result = await db.UpdateTenant(JsonConvert.DeserializeObject<TenantModel>(JsonConvert.SerializeObject(context)));
             if (result.Item1)
@@ -63,13 +72,44 @@ namespace Admin.Portal.API.Controllers
         [HttpDelete, Route("Delete")]
         public async Task<IActionResult> Delete(int id)
         {
+            if (!new AuthenticateFilter(config, dbContext).Authenticate(AccessLevel.Admins, Claims.CLAIM_READ_TENANT, Request.Headers[Config.HEADER_LOGIN_USER].ToString()))
+                return new Context(Messages.ACCOUNT_INSUFFICIENT_PRIVILEGES).ToContextResult((int)HttpStatusCode.Forbidden);
+
             IDataAccess db = ServiceInit.GetDataInstance(config, dbContext);
             bool result = await db.DeleteTenant(id);
             if (result)
-                return new Context(Messages.USER_DELETE_SUCCESS).ToContextResult();
+                return new Context(Messages.TENANT_DELETE_SUCCESS).ToContextResult();
             else
                 return new Context(Messages.DATA_ACCESS_FAILER).ToContextResult((int)HttpStatusCode.BadRequest);
         }
 
+        [HttpPut, Route("LinkRole")]
+        public async Task<IActionResult> LiknRole([FromBody] TenantRoleRequest context) 
+        {
+            if (!new AuthenticateFilter(config, dbContext).Authenticate(AccessLevel.Admins, Claims.CLAIM_READ_TENANT, Request.Headers[Config.HEADER_LOGIN_USER].ToString()))
+                return new Context(Messages.ACCOUNT_INSUFFICIENT_PRIVILEGES).ToContextResult((int)HttpStatusCode.Forbidden);
+
+            IDataAccess db = ServiceInit.GetDataInstance(config, dbContext);
+            (bool, TenantModel) result = await db.LinkRoleToTenant(context);
+            if (result.Item1)
+                return new Context((result.Item2)).ToContextResult();
+            else
+                return new Context(Messages.DATA_ACCESS_FAILER).ToContextResult((int)HttpStatusCode.BadRequest);
+        }
+
+        [HttpPut, Route("UnlinkRole")]
+        public async Task<IActionResult> UnLiknRole([FromBody] TenantRoleRequest context)
+        {
+            if (!new AuthenticateFilter(config, dbContext).Authenticate(AccessLevel.Admins, Claims.CLAIM_READ_TENANT, Request.Headers[Config.HEADER_LOGIN_USER].ToString()))
+                return new Context(Messages.ACCOUNT_INSUFFICIENT_PRIVILEGES).ToContextResult((int)HttpStatusCode.Forbidden);
+
+            IDataAccess db = ServiceInit.GetDataInstance(config, dbContext);
+            (bool, TenantModel) result = await db.UnLinkRoleFromTenant(context);
+            if (result.Item1)
+                return new Context((result.Item2)).ToContextResult();
+            else
+                return new Context(Messages.DATA_ACCESS_FAILER).ToContextResult((int)HttpStatusCode.BadRequest);
+        }
+      
     }
 }
